+++ src/components/cabecalho/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './styles.css'
+
+function Cabecalho(prop)
+{
+  return(
+    <div className="container-fluid">
+        <h1 className="title">Gera Nicks</h1>
+    </div>
+  )
+}
+
+export default Cabecalho;
