+++ src/components/rodape/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './styles.css';
+
+function Rodape(prop)
+{
+  return(
+    <div className = 'bg-dark text-white pt-3 footer'>
+      <h6>By: <a href="https://github.com/SrMeuNome">Sr Meu Nome</a></h6>
+    </div>
+  )
+}
+
+export default Rodape;
