+++ src/index.css
@@ -1,8 +1,12 @@
 body
 {
     background-color: #333;
-    background-image: url("./img/FundoPretoEAzul.jpg");
+    background-image: url("./assets/img/FundoPretoEAzul.jpg");
     background-blend-mode: color-dodge;
+}
+
+.container-fluid {
+    padding: 0;
 }
 
 .corpo
